文件下载前置逻辑实现（todo:文件下载存储操作相关逻辑实现）；

+++ cn.bossfriday.fileserver/src/main/java/cn/bossfriday/fileserver/engine/enums/FileStatus.java
@@ -2,16 +2,16 @@ package cn.bossfriday.fileserver.engine.enums;
 
 public enum FileStatus {
 
-    Normal((byte) 0),
-    Deleted((byte) 1);
+    Normal(0),
+    Deleted(1);
 
-    private final byte value;
+    private final int value;
 
-    FileStatus(byte value) {
+    FileStatus(int value) {
         this.value = value;
     }
 
-    public byte getValue() {
+    public int getValue() {
         return this.value;
     }
 }
